Store primary stats as floats

+++ src/common/mmo.hpp
@@ -590,7 +590,7 @@ struct mmo_charstatus {
 
 	char name[NAME_LENGTH];
 	unsigned int base_level,job_level;
-	unsigned short str,agi,vit,int_,dex,luk;
+       float str, agi, vit, int_, dex, luk;
 	unsigned short pow,sta,wis,spl,con,crt;
 	unsigned char slot,sex;
